Punto_4/requirements.txt
fastapi
uvicorn
Pillow
ultralytics
torch
